chore: rename Navbar → NavbarWrapper
