--- libs/core-wrappers/src/lib/Navbar/index.ts
@@ -1,2 +1,0 @@
-export { default as Navbar } from './Navbar';
-export { NavItem, QueryNavItem, AnchorNavItem } from './Navbar';
