+++ libs/core-wrappers/src/lib/NavbarWrapper/NavbarWrapper.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { NavLink } from 'react-router-dom';
 import { Icon } from '@tacc/core-components';
-import styles from './Navbar.module.css';
+import styles from './NavbarWrapper.module.css';
 
 type sizes = 'xs' | 'sm' | 'md' | 'lg';
